Testing Complete.
I've verified that Bounding Box Id 1 is being loaded with all the MapObjects present in it.  In turn each MapObject is being created and granted hardpoints and npc paths accordingly.

+++ database/createdb.sql
@@ -25,7 +25,7 @@ create table NonPlayerCharacterPath
 
 create table BoundingBox
 (
-   BoundingBoxId integer(10) primary key,
+   BoundingBoxId integer primary key,
    WC_X integer(10) null,
    WC_Y integer(10) null,
    Height integer(10) null,
